requirements.txt
# =================================================
# KRONOS DEFENSE V1.0 - DEPENDÊNCIAS DO SISTEMA
# =================================================

# --- INTERFACE E GRÁFICOS ---
customtkinter==5.2.2
Pillow==10.2.0
matplotlib==3.8.2

# --- INTELIGÊNCIA ARTIFICIAL (MOTOR ELITE) ---
opencv-python==4.9.0.80
face-recognition==1.3.0
numpy==1.26.4
ultralytics==8.1.0 
dlib==19.24.2

# --- INFRA, BANCO E COMUNICAÇÃO ---
psycopg2-binary==2.9.9
requests==2.31.0
python-dotenv==1.0.1
setuptools

# --- AUXILIARES DE PROCESSAMENTO ---
scipy==1.12.0
torch==2.2.0
torchvision==0.17.0